support for move up/down
This was surprisingly difficult! This completes the code for the set
dialog. I'll grep through for any more important TODO's -- maybe some
updates to the class code are in order with what I've learned from the
set code, as well -- and then I'll be moving on to the biggie, the
question editor. Once we can handle questions, sets, and classes, then
we can get started on generating quizzes, and finally on history.