save self.currentQid and update/create new based on that
This is a much more reliable way to handle it than what I was trying to
do before. I still wasn't immediately able to just update the question
name -- there is an earlier part of the onSave code that still relies on
the question name, so fixing that is the next priority.