implement deletion
This marks the end of all button implementations except for import and
export. Before I handle those, I'd like to work on consistency -- that
is, making sure it's not possible to end up with an unsaved question,
etc.

Two (hopefully) brief to-dos before that stage:
- Duplicate catching incorrectly ignores set number.
- Delete should bring up a confirmation window before wiping out a
  question.

Besides that, one other thing needs to be implemented in this dialog,
and that is the set change combo at the bottom. Also, the difficulty
spinner doesn't do anything, but since difficulty also isn't implemented
in the underlying db, we can't do anything about that yet.