disable the list part of the dialog when creating a new question
I'd like to also do this for editing a question, although that will
involve some fiddling to watch the appropriate focus events since we
don't have to explicitly click any buttons to edit a question.

Also add the "cancel new" button, which I'd like to change to "discard
changes" eventually once it can also apply to a question in editing.