quiz generation; fix Unicode error that came up suddenly
Unfortunately, PyRTF doesn't support Unicode unless it's encoded
already, so I had to add another library, rtfunicode. I don't understand
why this problem started suddenly, especially when the only characters
that got escaped after using rtfunicode were what appear to be ordinary
tabs. I do know that we need to support Unicode since this program is
going to be used for learning German with the expected set of non-ASCII
characters, so it's not a waste for that.

I'm also verrrrry much hoping that the TurningPoint import properly
supports Unicode, which is maybe suspect given its current setup, or
we're going to have very big problems, likely with extremely annoying
and progress-blocking workarounds.