connect the rest of the pieces
The essential elements of the program are basically finished, yay! The
only problem is that the scheduling algorithm is horribly broken --
that's my job tomorrow. I don't think it will require any significant
changes to the frontend code, though, so that's pretty much done. I
haven't, however, tested that everything works with properly scheduled
review cards.

A couple of UI tweaks need to be done still.