fix TurningPoint import issues
TurningPoint doesn't recognize Unicode-escaped tabs as tabs. I'm not
sure why rtfunicode feels the need to escape standard tab characters in
the first place, and I also bet that I wouldn't remember to handle that
case in my import code if I parsed an RTF file manually (another reason
to use plain text, dummies at TurningPoint), so this is squarely
rtfunicode's fault as far as I'm concerned. In any case, the fix is
happily simple and easy: just replace the code that it escapes to with a
tab character after running the encode.

Other Unicode characters that are necessary for German import properly
into TurningPoint after being escaped by rtfunicode, so we're cool there
as well.

Still to be done before delivery:
- fix the algorithm
- a couple of minor bug fixes that are written in PB3
- allow selection of the database location
- documentation, if I have the time?