allow resizing dialog boxes
This required significant redoing of the question edit dialog, since it
had been done kind of sloppily, without really properly nesting the
boxes. It took about an hour to get it back to the level it was at
previously, but now it resizes, which is very good.

I'd also like to add randomize/alphabetize all features to the button
set, and put a line between the single/all entry functions.