write autosave code and tests
Next is actually using autosave code instead of commits. The tests
definitely seem to indicate that the timer works correctly, but since I
can't close and reopen a database in RAM, I can't 100% confirm through
autotesting that the autosave trigger actually saves things. I doubt
there will be an issue, though.

Besides using checkAutosave() instead of a direct d.connection.commit(),
we need to make sure we close correctly. db.database.close() contains a
forceSave(), so this should not be a problem as long as the database is
closed when we close the program. We should also investigate what
happens if we SIGTERM it (do we have any error handling for that? I
don't think we do).