change all commits to autosaves
Wow. I feel very stupid. I did a lot of unnecessary work last year:
doing this instantaneously eliminated all performance issues and made
the program run perfectly snappy. Like the "QuestionManager" which holds
the state of questions in memory is really almost certainly completely
unnecessary. I'm not sure if it's worth redesigning to get rid of it at
this point, but it might be, because it's an unnecessary layer of
complexity (and it's possible that it actually slows down the
application somewhat with all the updates that it has to do and such).

I have to admit I'm not quite sure why so much committing was being done
when things were just being read from the database; that might be worth
looking into, actually, in case I coded something wrong in there. I do
know that that rather explains why things were running slowly, and that
it's definitely not necessary to do a complete save after every single
operation.