handle exceptions gracefully
Like Anki, some boilerplate and a traceback will be displayed and the
program's execution will continue if an error occurs. I think it's not a
good idea to force-save the database -- if the error was caused by the
database not working, it won't do any good, and if the program continues
to work afterwards, quitting will save just fine.