mark question list shortcut with underline
I'd not thought of this before, but while the text "Question List" can't
be buddied to the list, you can still use an ampersand to put the
underline there, and it looks like it works normally. This doesn't work
for the set list of course, but there is some similarity, and it's
probably less important there. I've retained the tooltip in the set list
but deleted it as redundant here.