add menus; implement File -> Open
A couple of other actions on the menus that I think would be useful,
like "show manual" and "back up database." Also, we can implement the
new-database-creation code within the program rather than in a separate
script that needs to be run.

We need to work out what to do when no database exists yet, though; a
new user would be really befuddled at being asked to select a database
on program start when she hadn't created any yet.