three small bugfixes
- For some reason, QSettings was often failing to correctly save the
last database used on quit -- the writeConf() call was running
correctly, but it wasn't actually getting written out to the config
file. Doing a manual sync() solved the problem, but I don't know why
this wasn't working, as it's supposed to get called by the destructor.

- On my system (or any Linux) it was possible to create a database that
didn't end in ".db", which I don't want to be easy since it's the
extension I'm using. (You can still rename it and then open it if you
*really* want it.) I guess maybe I could make my own extension, but it's
probably not really necessary. Maybe ask Mama what she would prefer.

- It was possible to open the generate quiz dialog with no classes,
causing an unhandled exception.