"generate quiz" and "randomize answers" buttons in question editor
I was going to do an "alphabetize questions" option, but realized that
would really be pretty pointless. Then I needed another button to put
next to "randomize answers" so there would be an even number, so I
thought up the generate-quiz option.

Also, added a vertical line between the buttons that act on the current
question (new, delete, move) and the ones that act on the whole list
(import, export, generate, randomize).

I *did* remember to add the new buttons to the list of elements that
need to get disabled when a question is being edited. Yay me!