this feature turned out to be a huge failure.
I don't think we need to do it, actually. It's kind of weird to have
error checking both in the db and in the ui, checking for different
things, but after trying for about an hour to make the stupid thing
work, I understand why I did it that way! There just are some things
that are better done when adding through the ui than in prevalidate(),
like checking if the question is a dupe (otherwise updating an existing
question fails! that's what pushed me over the edge). The only real
trick is to make sure that we can't get in an inconsistent state when
importing, either, which may be worth a bit of work still.