make things work
I decided I do need to go ahead with this branch in the end, although
(this is silly) at this point I don't remember why. I think there was
another issue with the original system and I figured I'd gotten so far
in on here that I should do it.

I've done a fair bit of testing, but I still need to do some more to
make sure every case of possible invalid question entry is covered. It
wouldn't hurt to write some more tests, but most of the problem is the
interaction between the ui and db, I think.

I also uncovered a bug that's been in there if not since the beginning,
for some time: if a question is edited to exactly match another
question, the discard button doesn't work correctly. I need to fix this
also.