latex rendering
Things that still need to be done on the backend side of this:

- We ought to borrow some of the 'texify' regex routines from RPPAS so
that we don't end up with misturned quotes, &c, on the quiz.

- We need to make a routine that actually runs xelatex (and allows later
editing); right now it creates a LaTeX string but doesn't actually write
it out to a file or generate a pdf. Some of this needs to be handled in
combination with the ui part, so it might be best to start on that
first.

- We could maybe also create some markup with ** as italic and such?
Also, we should make the [...] replacement and maybe the '//' one work
for the other quiz formats such as TP and preview.