add preference to autoselect answer choice A
We *really* need to redo the preferences framework -- it's such an
unbelievable pain to add a new preference, and there's a lot of
duplicated work and room for error. I think all the code that handles it
in ui/main should be moved into ui/prefs, and the preferences themselves
should be stored in a dictionary in the ConfigurationManager. Rather
than using writeConf/readConf externally, those should be private
methods and read/dump should be handled automatically when the program
starts and when the preferences dialog is accepted.

Also, there should be provision for initializing a QSettings store on a
new machine, rather than just relying on the fact that it can be handled
by whatever uses the preference.