warn if a lot of correct answers seem to be "A"
Warning is triggered when leaving the question editor: that is, when
closing the window or trying to generate a quiz (in the latter case, it
would be too late to warn when closing the window if the user had only
just created the questions and now wanted to immediately make a quiz).

It would be nice to also warn if a bunch were other letters, but that
would be getting really complicated for only a little gain -- A is the
only one that's really particularly likely to happen, I think.