add paper quiz output routine
Whoa! that's a lot of changes.

We still need to:
- Clean up the routine; for one, there should be a message indicating
that the save was successful, as that's not at all clear right now.
- Allow choosing whether or not you want to cancel when you hit the
logical cancel button (right now the utils.QuestionBox only gives you an
'OK' button for some reason).
- Code the HTML version of the quiz, and connect the text version (it
can use the same text as for the preview and just write it into a text
file).