an attempt at a bouncing progress bar
This proved waaaaaay harder than it should have been (for instance,
failing to close the dialog box in the code resulted in the X server
crashing with no traceback or log), and even now it's not quite the
right thing because the box takes too long to load up. I'm going to try
an hourglass instead now, but I just want to commit this so I can grab
the code I have here for something else later if appropriate.