begin implementing students window
A lot more work remains to be done; roughly in order of priority:

- Add/delete need to be implemented.
- The list should be sortable.
- A way to edit a student would be nice.
- This dialog should be accessible from the classes window.
- It would be nice to make it possible to copy a student in from another
  class. This is an optional extra and depends on editing being
  plausible.
- Import/export need to be implemented.

Before anything else can be done, though, we need to fix the DB schema:
I forgot that it's necessary for students to be associated with a
specific *class*, because they might have a different TP ID / etc. in a
different class, and it's anyway not useful to have students that aren't
in the class at issue showing up in the list.