make all cells editable; begin on adding and deleting
Adding works, but is kind of sloppy; I'd like to place the cursor in the
first cell immediately after clicking the add button, for instance.
Deleting deletes an item from the table but not from the database. In
both cases, the correct functions and calls are not really being used,
and the docs warn this could lead to "unexpected behavior," which is the
most annoying part of Qt, so I'll try to avoid that by changing over to
more official ways in a moment.

Also, we need to have some way of validating students, at the very least
when we close the window: two students probably shouldn't be allowed to
have the same first and last name (possible there would be a conflict,
but better to make the teacher do something about it by adding some
disambiguator to the field, I think) and definitely not the same TP ID
and device name. Email at least gets a warning on setting; maybe the
name too.