allow students dialog to be opened from classes
I also redesigned the classes dialog -- it made more sense to have the
buttons on the right, like in the sets dialog, than along the bottom.
That will make the interface look a little bit more consistent, and I
think it also saves square inches of screen space, as otherwise the
classes dialog would have to be unnecessarily wide to fit all the
buttons in.