implement sorting students; dialog tweaks
Sorting was surprisingly easy, once I got past the trap of directly
calling sort() in the model to set the initial sort, using the view's
function instead. (It does make a lot more sense to do it through the
view; that just wasn't the first function I saw!)

The tooltip explaining the editing functions of the table was getting in
the way constantly, so I disabled it and will just try to make sure it's
properly documented, though I have to say it seems rather unintuitive. I
don't have a better solution.