start using pylint and begin fixing complaints on db side
There are a lot of complaints to deal with. Genquiz is, I know, a total
mess, so I tried to fix some of the obvious things but didn't do any
refactoring for now -- we may not even keep the scheduling algorithm as
is, so it's not worth working on that right now.