begin trying some ideas for quiz history
- Make sure that my idea about pickling the question list would work (of
course, I didn't write it into a database, but when pickling and
unpickling from a string works, I'm sure I can get sqlite to hold the
string in between there).

- Otherwise check that the preview will render just fine without going
through methods of Quiz.

I've also made some paper notes that I haven't done anything about yet:
I'd like to get rid of the odd decorated methods in db/genquiz entirely.
I might also like to refactor Quiz so that more parameters are put in at
once; I think the problem is that the ui needs to be responsive to the
selected items, so it's useful to have the Quiz object built up over
time and return the data that it knows when it knows it. Just looking
over this interaction and writing in more comments/docstrings, and maybe
rearranging some of the methods in the physical source file, would
probably make it quite a bit easier to read/understand even if I don't
change any actual operational code.