bugfix: don't overwrite existing files on Linux
Because the Linux save dialog doesn't automatically add a file
extension, I had written a routine that would add one if so, but that
made it possible to overwrite a file that had the new name (e.g., if you
entered 'foo' in the filename box and there was a file named 'foo.rtf'
in the directory already, it would silently overwrite 'foo.rtf' with the
new content). This adds a utils routine to handle this properly.

Also, since I created an irritating bug by typing a constant as '16834'
instead of '16384', I switched to using the better-practice version of
'QMessageBox.Yes' in all placed I'd previously used it.