create email-results dialog in designer
We're going to have all of the email parameters exposed in the dialog
and use a format string syntax to enable the user to customize exactly
what shows up in the email (and we'll save the defaults from last time
the email function was used, of course). Something like this:

$c: the name of the current class
$n: the number of the current quiz
$s: name of the student this is being sent to
$a: list of student's answers vs. correct (like in view dialog)
$q: display of the quiz, with correct answers
$Q: display of the quiz, with correct & student's answers
$r: number right
$t: total number of questions
$p: percentage correct

Maybe others?