extensive work on emailing
Front-end password show/hide mostly implemented and secured in terms of
save/load. On the back end, get the format string parser working, and a
little bit of work on other infrastructure as well.

I will clean up the format-string parser's docstring in the next commit,
as I had been taking a bunch of notes there on what was required that I
want to save just in case I want them later, so they are preserved in
this commit.