continue work on emailing; change db access functions
Whoa! Implementing a progress bar has brought to light a number of
not-so-great things about my application; mostly, that the db
functionality isn't thread-safe at all, since there's only one
connection. That has not been fixed in this commit, but I think now that
queries are run through a method of a persistent class, it should be
relatively easy to add functionality to that class that can open a
second connection if it's necessary to allow a different thread to
access it.