finish emailing functions (error handling, etc.)
Some code cleanup could still be done, but I'm holding off for now:
there are better things to do and it's still possible bugs will be
discovered and the commented-out debug code &c will be useful.

Refactored sets to use the createNew() classmethod, as it was causing an
occasional database locking error to use the unnecessary-dump method.