Revert "steps toward non-modality"
This reverts commit eef6442d64fc2279c4d8db9183e55233a690d853.

Weird behavior was created by this; I assume it has something to do with
multithreading, even though I'm not explicitly creating multiple
threads. Being able to have it work nicely this way would be nice, but
it's not crucial at all, and it's definitely not worth trying to debug
seeing that it causes problems.

We can look back at this later if we want to try to make it work again.