latex server support
Note that no error-checking whatsoever has been done so far -- I will
add at least a couple of checks next.

A good future feature would be adding a progress bar (even just a
bouncing one) to the remote LaTeX generation function, as depending on
the server it may be as slow as molasses. This is what I get for using a
$35 computer, but on my RasPi it takes nearly a minute to compile,
retrieve the PDF, and clean up the temporary files, during which time
the application is just blocked with an hourglass shown!