don't start with remote option selected
By the way, it really doesn't make sense to have options in the
global preferences that are local to the db, especially when the radio
button is a global option and the ones underneath are local. I didn't
think very hard about this when I designed it -- I just thought it would
be good to guard the password a little better by putting it in the db.
This is something that should be corrected soon -- it should definitely
be a global option for logical reasons.